Crate testlib

Test library functions

These are common functions used in the integration tests for the script and for the lambda.

Structs

MetadataJson

Represents the fields to be anonymised

Functions

create_package

Creates a test tar.gz file

decompress_test_file

Decompresses the test tar.gz file

get_metadata_json_fields

Read the metadata.json file and parse the fields to be anonymised

json_missing_filename

An input string with the filename missing

metadata_from_json_value 🔒

Parse the fields from the json value

valid_json

An valid input string